// src/types/contents.tsx
export interface IContent {
  id: number;
  createdAt: string;
  updatedAt: string;
  isArchive: boolean;
  name: string;
  surname: string;
  nickname: string;
  userId: string;
  img: string;
  nationality: string;
  ageLastSeen: number;
  dateOfBirth: string;
  missingDetail: string;
  gender: string;
  weight: number;
  height: number;
  skin: string;
  remark: string;
  status: string;
  province: string;
  place: string;
  missingDatetime: string;
}